docs(Example Modules, Module Styles): add code example of the output style value `array` format.
In this commit, we add code example of the output style value in `array` format. The code is intentionally added and commented in FE only as an example of expected value format. It's added to `module_styles` of all example modules:
- ChildModule
- D4Module
- DynamicModule
- ParentModule
- StaticModule

If 3P devs have custom style processing, the style output should be passed as an `array` of style declarations
to the `styles` property of the `Style::add` method. For example:
```
[
    [
        'atRules'     => false,
        'selector'    => $icon_selector,
        'declaration' => 'color: red;'
    ],
    [
        'atRules'     => '@media only screen and (max-width: 767px)',
        'selector'    => $icon_selector,
        'declaration' => 'color: green;'
    ],
],
```

Part of https://github.com/elegantthemes/Divi/issues/40336.