fix(D5, Example Modules, Placeholder Content): use `@divi/module-utils` to import `placeholderContent`.
Recently, we just move the `placeholderContent` from `module` to `module-utils` package. This causes a fatal error on console log when it's being used by example and third-party modules. This issue happens due to the `divi-types` is not updated yet, so the `placeholderContent` is still in `module` package instead of `module-utils` package.

Despite we can simply update the import, the error still exists on build when we run `npm run start`. So, we need to update our `divi-types` fully to make sure we can fix the issue. For note, I added quick note on readme file to make anyone who use this plugin know the error.

Part of elegantthemes/Divi#42110.