fix(Static Module, Summary Field): Remove placeholder content from initial value.
### What is the problem?

I forgot to remove the placeholder content from the Summary field initial value. It shouldn't be there.

### Why does the problem happen?

The placeholder value on the field and placeholder content are 2 different concept.

**Placeholder value** is the default value that we want to display on the **field setting** when the field is "empty". For example, if you go to Title Text Size field option, the `26px` there is default value and displayed as placeholder of the input field.

**Placeholder content** is basically the initial value that will be used when the module is initially added. For example, if you add Audio module, "Your Title Goes Here" text is automatically used on Title text. Here, the "Your Title Goes Here" text is field value, not placeholder (default) value. User can enable disable this behavior by disabling "Add Placeholder Content To New Modules" on the Builder settings.

So, we just need to define the "default value a.k.a. placeholder value" here, not "placeholder content". So, if I remove this one it should works as I explained below.

Right now the Summary field has "This is a brief summary that provides an overview of the content below. It can span multiple lines to give readers a comprehensive understanding of what to expect in the main content section." text as default value that will be displayed as placeholder text.

### How the fix works

The goal of this PR is to display the placeholder value on the **Summary field on Static Module settings**  when the value is "empty":
- Visually or user perspective, the "empty" here means there is no value set on the field.
- Technically, the "empty" here means either:
  - The value is set as empty string
  - The value is reset which means we're back to use the default value (not just empty string).

**Illustration of set empty string**

The "value" is just empty string "".

```
attrs: {
  "summary": {
  "innerContent": {
    "desktop": {
      "value": ""
    }
  }
},
```

**Illustration of reset field (remove attribute)**

The "value" is back to default value.

```
attrs: {
  "summary": {
  "innerContent": {
    "desktop": {
      "value": "This is a brief summary that provides an overview of the content below. It can span multiple lines to give readers a comprehensive understanding of what to expect in the main content section."
    }
  }
},
```

Here the latest video:

https://github.com/user-attachments/assets/cca08644-4e0e-41fd-9873-5683a119c225

In the video above, I do some steps:

**Add Static Module**

When I add the Static module, the "default value" is displayed as "placeholder value" on the Text area of Summary field. The rendered module also display the "default value" in the VB layout.

**Reset Field Value (Back to Default)**

Next, I set certain value, then I reset the field. Since reset means back to default value, now the Text area field shows the "default value" as placeholder again and the rendered module also display the "default value" in the VB layout, just like when we add the Static module.

**Delete Field Value (Empty String)**

Finally, I set certain value again, but this time I will delete the value on the field. Since delete means we will set the value as empty string (not reset to use the default value), the Text area field will also show the "default value" as placeholder again, but the rendered module will display the empty string in the VB layout.

Why this happen? Because the `showPlaceholderOnEmpty` is a property of the Text area component to show default value as placeholder when the field is empty. This property doesn't have any correlation with the rendered element. So, this is expected behavior.

How if 3P devs want to change the behavior where they want to display? They can simply modify the module attribute like this:

```
  // When summary is reset, placeholder is displayed since reset restores default value.
  // When summary is empty, placeholder is hidden since empty string indicates intentional user action.
  // To modify this behavior, developers can update the summary attribute with a default value during module render.
  if ('' === attrs?.summary?.innerContent?.desktop?.value) {
    elements.attrs = {
      ...elements.attrs,
      summary: {
        ...elements.attrs.summary,
        innerContent: {
          desktop: {
            value: staticMetadata?.attributes?.summary?.default?.innerContent?.desktop?.value ?? '',
          },
        },
      },
    };
  }
```

The code above will make the rendered module layout display the default value when the value is empty.

https://github.com/user-attachments/assets/48a4a9ce-74f2-41cd-9fc4-8e2c2778156e

However, since that is not D5 default behavior on how empty string works, the code will be added only as comment in case 3P devs want to use the same approach.